==> cluster-analysis/notebooks/dataset/1-icecat_taxonomy.ipynb <==
# Icecat Taxonomy

Icecat defines a taxonomy per product category. The dataset should contain products from the category "Mobile Phone Cases" and "Smartphones". So, a look on the Icecat Taxonomy for both categories has to be taken now.

Both taxonomies were downloaded from Icecat as Excel files to the data directory.


import pandas as pd

from src import config

## Loading the Taxonomy Files


dir_data_src = config.dir_data / "icecat-taxonomy"
dir_data_result = config.dir_data / "dataset"

taxonomy_cases = pd.read_excel(dir_data_src / "mobile_phone_cases.xlsx")
taxonomy_cases

taxonomy_phones = pd.read_excel(dir_data_src / "smartphones.xlsx")
taxonomy_phones

## Analysis of the Columns


pd.merge(taxonomy_cases, taxonomy_phones, how="outer", on="Feature_id")

Merge all the different columns to see, if they are uniform for the two product categories:

cols = ["Feature", "Feature group", "Type", "Restricted values", "Measure sign"]
num_of_entries = []

for col in cols:
    merged_df = pd.merge(taxonomy_cases, taxonomy_phones, "outer", ["Feature_id", col])
    num_of_entries.append(len(merged_df))

pd.DataFrame({"merge on": cols, "num of entries": num_of_entries})


Comparing the feature groups with each other. The groups for the "phones" category seem to be more sophisticated. So, we will use the grouping of the "phones" category:

cols = ["Feature_id", "Feature", "Feature group"]

merged_df = pd.merge(
    taxonomy_cases[cols],
    taxonomy_phones[cols],
    "outer",
    ["Feature_id", "Feature"],
    suffixes=["_cases", "_phones"],
).dropna()

merged_df[merged_df["Feature group_cases"] != merged_df["Feature group_phones"]]

The data types only differ for the feature 8072. Both are drop downs. But for "phones" it is a multi dropdown. We will also use the "phone"'s type in this case:

cols = ["Feature_id", "Feature", "Type"]

merged_df = pd.merge(
    taxonomy_cases[cols],
    taxonomy_phones[cols],
    "outer",
    ["Feature_id", "Feature"],
    suffixes=["_cases", "_phones"],
).dropna()

merged_df[merged_df["Type_cases"] != merged_df["Type_phones"]]

## Compile Result


### Attribute Groups

groups_df = pd.merge(
    taxonomy_cases["Feature group"],
    taxonomy_phones["Feature group"],
    "outer",
).drop_duplicates().rename({"Feature group": "locale_en"}, axis=1)

groups_df["code"] = (
    groups_df["locale_en"].str.replace("&", "and").str.replace(" ", "_").str.lower()
)

groups_df = groups_df.sort_index(axis=1).sort_values("code")
groups_df

Attribute groups are now stored to a csv file. Afterwards, the german localizations are added manually.

groups_df.to_csv(dir_data_result / "attribute-groups.csv", index=False)

### Attributes & Families

cols = ["Feature_id", "Feature", "Feature group", "Type", "Is mandatory"]

merged_df = pd.merge(
    taxonomy_cases[cols],
    taxonomy_phones[cols],
    "outer",
    cols[0:-1],
).drop_duplicates(cols[0], keep="last") # prefer group and type from `taxonomy_phones`

merged_df

(merged_df["Feature_id"] < 2).drop_duplicates()

map_col_names = {
    "Feature_id": "code",
    "Feature": "locale_en",
    "Feature group": "group",
    "Type": "type",
    "Is mandatory_x": "mobile_phone_cases",
    "Is mandatory_y": "smartphones",
}
map_values = {
    0.0: "optional",
    1.0: "required",
}

renamed_df = merged_df.rename(map_col_names, axis=1).replace(map_values)
renamed_df["group"] = (
    renamed_df["group"].str.replace("&", "and").str.replace(" ", "_").str.lower()
)

attributes_df = renamed_df.sort_values("code")
attributes_df

Saving the Icecat Attributes to csv for later insert into Akeneo PIM:

attributes_df.to_csv(dir_data_result / "attributes-icecat.csv", index=False)
