==> cluster-analysis/notebooks/dataset/2-add-basics.ipynb <==
# Add Basic Settings to Akeneo PIM

import pandas as pd

from src import akeneo, config

client = akeneo.create_client_from_env()

## Base Category Tree

category_base = "master"

client.request("pim_api_category_partial_update", {"code": category_base}, body={
    "labels": {
        "en_US": "Master",
        "en_GB": "Master",
        "de_DE": "Master",
    }
})

## Default Channel with Locales and Currencies

channel = "default"

client.request("pim_api_channel_partial_update", {"code": channel}, body={
    "category_tree": category_base,
    "currencies": ["USD", "EUR"],
    "locales": ["en_US", "en_GB", "de_DE"],
    "labels": {
        "en_US": "Default",
        "en_GB": "Default",
        "de_DE": "Default",
    }
})

## Base Attributes for Icecat Connector

### Attribute Group

Basic attribute group for all icecat base attributes:

attr_group = "basic"

client.request(
    "pim_api_attribute_group_partial_update",
    {"code": attr_group},
    {
        "labels": {
            "en_US": "Basic",
            "en_GB": "Basic",
            "de_DE": "Basic",
        },
    },
)

Attribute group for faulty attributes, that the icecat importer does not handle correctly:

client.request(
    "pim_api_attribute_group_partial_update",
    {"code": "faulty"},
    {
        "labels": {
            "en_US": "Faulty",
            "en_GB": "Faulty",
            "de_DE": "Fehlerhaft",
        },
    },
)

### Load Basic Attributes from CSV File

attr_file = config.dir_data / "dataset" / "attributes-basic.csv"
attr_df = pd.read_csv(attr_file)
attr_df

for _, code, type_, locale_en, locale_de, opts_str in attr_df.itertuples():
    res = client.request("pim_api_attribute_partial_update", {"code": code}, {
        "type": type_,
        "labels": {
            "en_US": locale_en,
            "en_GB": locale_en,
            "de_DE": locale_de,
        },
        "group": attr_group,
        **eval(opts_str),
    })

    print(code, "=>", res)
