==> cluster-analysis/notebooks/dataset/3-fix-icecat-attributes.ipynb <==
# Check Import of Icecat Attributes

The icecat attributes were entered by hand into Akeneo. In this notebook we check, if that process went well.

import pandas as pd

from src import akeneo, config

## Getting Data

data_dir = config.dir_data / "dataset"

### Load Wanted Attributes

attributes_want_df = pd.read_csv(data_dir / "attributes-icecat.csv")
attributes_want_df["code"] = attributes_want_df["code"].map(lambda x: f"icecat_{x}")
attributes_want_df

### Get Attributes in PIM

client = akeneo.create_client_from_env()

attributes_got_all = client.request("pim_api_attribute_list")
attributes_got_all_df = pd.DataFrame(attributes_got_all)
attributes_got_all_df

Remove all basic and fixed attributes

attributes_got_df = attributes_got_all_df[
    (attributes_got_all_df["group"] != "basic") &
    ~attributes_got_all_df["code"].str.contains("_fixed")
]
attributes_got_df

## Checking Attributes

### Not Imported

Some attributes could not be imported into Akeneo for an unknown reason.

attributes_want_df[
    ~attributes_want_df["code"].isin(attributes_got_df["code"])
]

### Unnecessary Imports 

attributes_got_df[
    ~attributes_got_df["code"].isin(attributes_want_df["code"])
]

### Check Type Mapping

The Icecat types and Akeneo PIM types differ. Here we take a look, how they are converted into each other.

tmp_got = attributes_got_df[["code", "type"]]
tmp_want = attributes_want_df[["code", "type"]].copy()

merged_df = pd.merge(tmp_got, tmp_want, how="outer", on="code", suffixes=["_akeneo", "_icecat"])
merged_df

#### Check all type mappings

There are several mappings, that did not go well. They cannot be changed in Akeneo, because the Icecat Importer does not work properly then. So they have to be fixed afterwards.

cols = ["code", "type_akeneo", "type_icecat"]
merged_df[cols].groupby(cols[1:]).count()

preparing attribute fixes:

attr_to_fix_df = merged_df[
    (merged_df["type_akeneo"] == "pim_catalog_simpleselect")
    & (merged_df["type_icecat"] == "multi_dropdown")
    | (merged_df["type_akeneo"] == "pim_catalog_text")
    & (merged_df["type_icecat"] == "dropdown")
    | (merged_df["type_akeneo"] == "pim_catalog_text")
    & (merged_df["type_icecat"] == "numerical")
].sort_values("code")

additional_attr = pd.DataFrame([
    {"code": "icecat_brand", "type_akeneo": "pim_catalog_text", "type_icecat": "dropdown"}
])

attr_to_fix_df = pd.concat([attr_to_fix_df, additional_attr])

map_icecate_to_akeneo_type = {
    "multi_dropdown": "pim_catalog_multiselect",
    "dropdown": "pim_catalog_simpleselect",
    "numerical": "pim_catalog_number",
}

attr_to_fix_df["target_type"] = attr_to_fix_df["type_icecat"].map(map_icecate_to_akeneo_type)

attr_to_fix_df

attr_to_fix_df.to_csv(data_dir / "attributes-fixes.csv", index=False)

Create attributes for fixing:

responses = []

for _, attr_code, target_type in attr_to_fix_df[["code", "target_type"]].itertuples():
    attr = client.request("pim_api_attribute_get", {"code": attr_code})
    del attr["code"]
    del attr["group_labels"]

    if target_type == "pim_catalog_number":
        attr["decimals_allowed"] = True
        attr["negative_allowed"] = True

    attr_code_fix = f"{attr_code}_fixed"
    res = client.request(
        "pim_api_attribute_partial_update",
        {"code": attr_code_fix},
        {**attr, "type": target_type},
    )
    responses.append({"attr": attr_code_fix, **res})

df = pd.DataFrame(responses)
df

### Check Locales

attr = client.request("pim_api_attribute_list")
attr_df = pd.DataFrame(attr)

tmp = []

for _, code, labels in attr_df[["code", "labels"]].itertuples():
    tmp.append({"code": code, **labels})

locales_df = pd.DataFrame(tmp)

Check for missing locales:

locales_missing_df = locales_df[
    (locales_df["en_US"] == "") |
    (locales_df["en_GB"] == "") |
    (locales_df["de_DE"] == "")
]
locales_missing_df
