==> cluster-analysis/notebooks/dataset/4-structure-icecat-attributes.ipynb <==
# Structuring Imported Icecat Attributes

After the successful import of the Icecat attributes, we now have to assign them to their correct attribute group. Also, we prepare the product families with the correct attribute requirements.

import pandas as pd

from src import akeneo, config

## Load Attributes Data

data_src = config.dir_data / "dataset"

attr_basic_df = pd.read_csv(data_src / "attributes-basic.csv")
attr_basic_df

attr_icecat_df = pd.read_csv(data_src / "attributes-icecat.csv")
attr_icecat_df["code"] = attr_icecat_df["code"].map(lambda x: f"icecat_{x}")
attr_icecat_df

attr_fixes_df = pd.read_csv(data_src / "attributes-fixes.csv")
attr_fixes_df

attr_groups_df = pd.read_csv(data_src / "attribute-groups.csv")
attr_groups_df

## Structure Icecat Attributes

client = akeneo.create_client_from_env()

Remove all Icecat attributes from the list, that could not be imported:

attr_in_pim = client.request("pim_api_attribute_list")
attr_codes_in_pim = pd.DataFrame(attr_in_pim)["code"]
attr_icecat_df = attr_icecat_df[
    attr_icecat_df["code"].isin(attr_codes_in_pim)
]
attr_icecat_df

### Attribute Groups

group_faulty = "faulty"
map_group_to_attr_codes: dict[str, list[str]] = {
    group_faulty: [],
}

attr_codes_to_fix = attr_fixes_df["code"].to_list()

for _, attr_code, group in attr_icecat_df[["code", "group"]].itertuples():
    if group not in map_group_to_attr_codes:
        map_group_to_attr_codes[group] = []

    if attr_code in attr_codes_to_fix:
        map_group_to_attr_codes[group_faulty].append(attr_code)
        map_group_to_attr_codes[group].append(f"{attr_code}_fixed")
    else:
        map_group_to_attr_codes[group].append(attr_code)

pd.DataFrame(map_group_to_attr_codes.items())

for _, code, en, de in attr_groups_df.itertuples():
    attributes = map_group_to_attr_codes.get(code, [])
    res = client.request("pim_api_attribute_group_partial_update", {"code": code}, {
        "attributes": attributes,
        "labels": {
            "en_US": en,
            "en_GB": en,
            "de_DE": de,
        }
    })
    print(code, "=>", res)

### Families

family_attributes = {
    "mobile_phone_cases": {
        "optional": [],
        "required": [],
    },
    "smartphones": {
        "optional": [],
        "required": [],
    },
}

attr_codes_to_fix = attr_fixes_df["code"].to_list()

for _, attr_code, cases, phones in attr_icecat_df[["code", "mobile_phone_cases", "smartphones"]].itertuples():
    if cases in ["optional", "required"]:
        family_attributes["mobile_phone_cases"][cases].append(attr_code)
        if attr_code in attr_codes_to_fix:
            family_attributes["mobile_phone_cases"][cases].append(f"{attr_code}_fixed")

    if phones in ["optional", "required"]:
        family_attributes["smartphones"][phones].append(attr_code)
        if attr_code in attr_codes_to_fix:
            family_attributes["smartphones"][phones].append(f"{attr_code}_fixed")

pd.DataFrame(family_attributes)

attr_basic = attr_basic_df["code"].to_list()

families = [
    ("mobile_phone_cases", "Mobile Phone Cases", "Smartphone Hüllen"),
    ("smartphones", "Smartphones", "Smartphones"),
]

for family_code, en, de in families:
    attr_optional = family_attributes[family_code]["optional"]
    attr_required = family_attributes[family_code]["required"]

    res = client.request("pim_api_family_partial_update", {"code": family_code}, {
        "attribute_as_label": "icecat_title",
        "attribute_as_image": "icecat_image_0",
        "attributes": [*attr_basic, *attr_required, *attr_optional],
        "attribute_requirements": {
            "default": [*attr_basic, *attr_required],
        },
        "labels": {
            "en_US": en,
            "en_GB": en,
            "de_DE": de,
        }
    })

    print(family_code, "=>", res)
