==> cluster-analysis/notebooks/dataset/5-add-products.ipynb <==
# Adding the Products

import pandas as pd

from src import akeneo, config

data_dir = config.dir_data / "dataset"

client = akeneo.create_client_from_env()

## Categories

### Load Categories from File

categories_df = pd.read_csv(data_dir / "categories.csv")
categories_df

### Insert Categories into PIM

for _, parent, code, en, de in categories_df.itertuples():
    res = client.request(
        "pim_api_category_partial_update",
        {"code": code},
        {
            "parent": parent if pd.notna(parent) else None,
            "labels": {
                "en_US": en,
                "en_GB": en,
                "de_DE": de,
            },
        },
    )
    print(code, "=>", res)

## Products

### Load Products from File

products_file_df = pd.read_csv(data_dir / "products.csv")
products_file_df

products_file_df.describe()

### Get SKU and EAN from Icecat

def get_sku_and_ean(url: str):
    ean_parts = url.split("/")[7].split("-")
    ean = ""
    for part in ean_parts:
        if part.isnumeric():
            ean = part
            break

    return (
        url,
        url.split("/")[6].upper(),
        ean,
    )

print(get_sku_and_ean("https://icecat.biz/en/p/samsung/ef-pg980tbegeu/mobile+phone+cases-8806090267246-ef-pg980-77657600.html"))
print(get_sku_and_ean("https://icecat.biz/de/p/samsung/sm-s901bzadeue/galaxy-smartphones-8806094306170-sm-s901b-ds-97302928.html"))

sku_and_ean_df = pd.DataFrame.from_records(products_file_df["url"].map(get_sku_and_ean), columns=["url", "sku", "ean"])

products_df = pd.merge(
    products_file_df,
    sku_and_ean_df,
    on="url"
)
products_df["categories"] = products_df["categories"].str.split(",")
products_df

### Insert Products into PIM

responses = []
for _, family, brand, categories, url, sku, ean in products_df.itertuples():
    res = client.request(
        "pim_api_product_partial_update",
        {"code": sku},
        {
            "family": family,
            "categories": categories,
            "values": {
                "ean": [{"locale": None, "scope": None, "data": ean}],
                "icecat_brand": [
                    {"locale": "en_US", "scope": "default", "data": brand},
                    {"locale": "en_GB", "scope": "default", "data": brand},
                    {"locale": "de_DE", "scope": "default", "data": brand},
                ],
            },
        },
    )
    responses.append({"sku": sku, **res})

df = pd.DataFrame(responses)
df.head()

df[~df["status"].isin([201, 204])]

### Check for Not Imported Products

products_in_pim_df = pd.DataFrame(client.request("pim_api_product_list"))
products_df[
    ~products_df["sku"].isin(products_in_pim_df["identifier"])
]
