==> cluster-analysis/notebooks/dataset/6-fix-product-values.ipynb <==
# Fixing Attributes and Product Values



import pandas as pd

from src import akeneo, config

## Loading Resources

fixes_df = pd.read_csv(config.dir_data / "dataset" / "attributes-fixes.csv")
fixes_df

client = akeneo.create_client_from_env()

## Fixing Values

def extract_locales(values: list[dict]) -> dict:
    result = {}
    for value in values:
        locale = value["locale"]
        data = value["data"]
        if locale not in result:
            result[locale] = data
    return result

def to_code(label: str) -> str:
    return (
        label.lower()
        .replace("(", "")
        .replace(")", "")
        .replace("-", "_")
        .replace(".", "_")
        .replace("@", "_at_")
        .replace("+", "plus")
        .strip()
        .replace(" ", "_")
    )

### Fix Numbers

def fix_number(attr_code: str) -> list[str]:
    logs = []

    attr_code_fix = f"{attr_code}_fixed"
    products = client.request("pim_api_product_list", params={"attributes": attr_code})
    for product in products:
        if attr_code not in product["values"]: continue

        prod_id = product["identifier"]
        values = product["values"][attr_code]

        locales = extract_locales(values)
        num = float(locales["en_US"].strip())
        
        for i in range(len(values)):
            values[i]["data"] = num
            
        res = client.request(
            "pim_api_product_partial_update",
            {"code": prod_id},
            {"values": {attr_code_fix: values}},
        )
        logs.append(f"update product {prod_id} => {res}")

    return logs

logs_numbers = []

numbers = fixes_df[fixes_df["target_type"] == "pim_catalog_number"]
for attr_code in numbers["code"]:
    log = fix_number(attr_code)
    logs_numbers.append(log)

### Fix Single Selects

def fix_single_select(attr_code: str) -> list[str]:
    logs = []

    options: set[str] = set()
    attr_code_fix = f"{attr_code}_fixed"

    products = client.request("pim_api_product_list", params={"attributes": attr_code})
    for product in products:
        if attr_code not in product["values"]: continue

        prod_id = product["identifier"]
        values = product["values"][attr_code]

        locales = extract_locales(values)
        opt_code = to_code(locales["en_US"])

        if opt_code not in options:
            res = client.request(
                "pim_api_attribute_option_partial_update",
                {"attributeCode": attr_code_fix, "code": opt_code},
                {"labels": locales},
            )
            logs.append(f"Add {opt_code} to {attr_code_fix} => {res}")
            options.add(opt_code)

        for i in range(len(values)):
            values[i]["data"] = opt_code
        res = client.request(
            "pim_api_product_partial_update",
            {"code": prod_id},
            {"values": {attr_code_fix: values}},
        )
        logs.append(f"update product {prod_id} => {res}")

    return logs

logs_single_selects = []

simple_selects = fixes_df[fixes_df["target_type"] == "pim_catalog_simpleselect"]
for attr_code in simple_selects["code"]:
    log = fix_single_select(attr_code)
    logs_single_selects.append(log)

### Fix Multi Selects

def extract_multi_options(attr_code: str):
    map_options: dict[str, list[str]] = {}
    options: dict[str, dict] = {}

    attr_opts = client.request(
        "pim_api_attribute_option_list", {"attributeCode": attr_code}
    )
    for attr_opt in attr_opts:
        code = attr_opt["code"]
        opts = list(map(to_code, (
            attr_opt["labels"]["en_US"].split(",")
        )))
        ens = attr_opt["labels"]["en_US"].split(",")
        des = attr_opt["labels"]["de_DE"].split(",")
        for i in range(len(opts)):
            if opts[i] not in options:
                options[opts[i]] = {
                    "labels": {
                        "en_US": ens[i].strip(),
                        "en_GB": ens[i].strip(),
                        "de_DE": des[i].strip(),
                    }
                }
        map_options[code] = opts

    return map_options, options

extract_multi_options("icecat_898")

def add_multi_options(attr_code: str, options: dict[str, dict]) -> list[str]:
    logs = []
    for opt_code, body in options.items():
        res = client.request(
            "pim_api_attribute_option_partial_update",
            {"attributeCode": attr_code, "code": opt_code},
            body,
        )
        logs.append(f"Add {opt_code} to {attr_code} => {res}")
    return logs

def add_multi_options_to_products(attr_code: str, map_options: dict[str, list[str]]) -> list[str]:
    logs = []
    products = client.request("pim_api_product_list", params={"attributes": attr_code})
    for product in products:
        prod_id = product["identifier"]
        if attr_code in product["values"]:
            values = product["values"][attr_code]
            for i in range(len(values)):
                values[i]["data"] = map_options[values[i]["data"]]
            res = client.request(
                "pim_api_product_partial_update",
                {"code": prod_id},
                {"values": {f"{attr_code}_fixed": values}},
            )
            logs.append(f"update product {prod_id} => {res}")
    return logs

def fix_multi_select(attr_code: str) -> list[str]:
    logs = []
    
    mapping, options = extract_multi_options(attr_code)
    logs.append(f"mapping: {mapping}, options: {options}")

    logs += add_multi_options(f"{attr_code}_fixed", options)
    logs += add_multi_options_to_products(attr_code, mapping)

    return logs

logs_multi_selects = []

multi_selects = fixes_df[fixes_df["target_type"] == "pim_catalog_multiselect"]
for attr_code in multi_selects["code"]:
    log = fix_multi_select(attr_code)
    logs_multi_selects.append(log)
